Update Trial tag to v1beta1-45c5727 in SDK examples (#1481)
* Update Trial image tag to v1beta1-45c5727 in SDK examples

* Move SDK examples

* Check if Experiment is running after the restart in e2e

+++ test/e2e/v1beta1/run-e2e-experiment.go
@@ -130,8 +130,9 @@ func main() {
 			}
 			time.Sleep(1 * time.Second)
 		}
-		if !exp.IsRestarting() {
-			log.Fatalf("Unable to restart Experiment %v", exp.Name)
+		// Check if Experiment is not restarting and is not running.
+		if !exp.IsRestarting() && !exp.IsRunning() {
+			log.Fatalf("Unable to restart Experiment %v, Experiment conditions: %v", exp.Name, exp.Status.Conditions)
 			// Delete experiment in case of error.
 			log.Printf("Deleting Experiment %v\n", exp.Name)
 			if kclient.DeleteRuntimeObject(exp) != nil {
